rwa_package/requirements.txt
pandas==1.0.2
pytest==5.4.1
numpy==1.18.2
pathlib==1.0.1
google.cloud==0.34.0